Made the login form, but not connected to API. Renamed files to match Angular's Official Style Guide
