+++ Angular/src/app/MainApp.js
@@ -1,0 +1,3 @@
+export const MainApp = {
+  templateUrl: 'app/MainApp.html'
+};
