--- Angular/src/app/app.js
@@ -1,3 +1,0 @@
-export const MainApp = {
-  templateUrl: 'app/app.html'
-};
