+++ Angular/src/app/components/Login/Login.html
@@ -1,4 +1,26 @@
 <div class="component" >
-  <h1>LOGIN COMPONENT</h1>
+  <h1>Login</h1>
+  <h2>{{ text }}</h2>
+
+  <form>
+    <label for="username">Username: </label>
+    <input type="text" name="username" id="username" value="" placeholder="Username" >
+
+    <!-- ============================= -->
+    <!-- Remove after adding bootstrap -->
+    <!-- ============================= -->
+    </br>
+
+    <label for="email">Email: </label>
+    <input type="email" name="email" id="email" value="" placeholder="Email" >
+    <!-- ============================= -->
+    <!-- Remove after adding bootstrap -->
+    <!-- ============================= -->
+    </br>
+
+    <label for="password">Password: </label>
+    <input type="password" name="password" id="password" value="" placeholder="Password" >
+  </form>
+
   <a ui-sref="app" ><button>Back Home</button></a>
 </div>
